Revert "fix(checker): preserve narrowing context across flow boundary"
This reverts commit 934b2a59a0a59d8a9d96e4e645c308f52583fc05.
